fixed build scripts

+++ frontend/build-docker.sh
@@ -4,8 +4,12 @@
 
 npm run build
 
-mkdir -p /usr/frontend/build
-cp -r dist/ /usr/frontend/build
+# If in a container, copy to volumes
 
-mkdir -p /usr/frontend/images
-cp -r images/ /usr/frontend/images
+if [ -f /.dockerenv ]; then
+    mkdir -p /usr/frontend/build
+    cp -r dist/ /usr/frontend/build
+
+    mkdir -p /usr/frontend/images
+    cp -r images/ /usr/frontend
+fi
